split builders and guards for maintainability
